Conflicts:
	apps/devApps/iosES2RendererExample/src/main.mm
	examples/android/androidAdvanced3DExample/.cproject
	examples/android/androidAssimpExample/.cproject
	examples/android/androidAudioExample/.cproject
	examples/android/androidCameraExample/.cproject
	examples/android/androidEmptyExample/.cproject
	examples/android/androidFontExample/.cproject
	examples/android/androidImageExample/.cproject
	examples/android/androidOpenCVExample/.cproject
	examples/android/androidOpenCVFaceExample/.cproject
	examples/android/androidPolygonExample/.cproject
	examples/android/androidPolygonExample/src/testApp.cpp
	examples/android/androidSoundPlayerExample/.cproject
	examples/android/androidSwipeExample/.cproject
	examples/android/androidTouchExample/.cproject
	examples/android/androidVBOExample/.cproject
	examples/events/eventsExample/src/testApp.cpp
	libs/openFrameworks/3d/ofCamera.cpp
	libs/openFrameworks/communication/ofSerial.cpp
	libs/openFrameworks/gl/ofProgrammableGLRenderer.cpp

+++ addons/ofxAssimpModelLoader/src/ofxAssimpModelLoader.cpp
@@ -307,7 +307,7 @@ void ofxAssimpModelLoader::update() {
 }
 
 void ofxAssimpModelLoader::updateAnimations() {
-    for(int i=0; i<animations.size(); i++) {
+    for(unsigned int i=0; i<animations.size(); i++) {
         animations[i].update();
     }
 }
@@ -322,26 +322,26 @@ void ofxAssimpModelLoader::updateMeshes(aiNode * node, ofMatrix4x4 parentMatrix)
                        m.d1, m.d2, m.d3, m.d4);
     matrix *= parentMatrix;
     
-    for(int i=0; i<node->mNumMeshes; i++) {
+    for(unsigned int i=0; i<node->mNumMeshes; i++) {
         int meshIndex = node->mMeshes[i];
         ofxAssimpMeshHelper & mesh = modelMeshes[meshIndex];
         mesh.matrix = matrix;
     }
     
-    for(int i=0; i<node->mNumChildren; i++) {
+    for(unsigned int i=0; i<node->mNumChildren; i++) {
         updateMeshes(node->mChildren[i], matrix);
     }
 }
 
 void ofxAssimpModelLoader::updateBones() {
     // update mesh position for the animation
-	for(int i=0; i<modelMeshes.size(); ++i) {
+	for(unsigned int i=0; i<modelMeshes.size(); ++i) {
 		// current mesh we are introspecting
 		const aiMesh* mesh = modelMeshes[i].mesh;
         
 		// calculate bone matrices
 		vector<aiMatrix4x4> boneMatrices(mesh->mNumBones);
-		for(int a=0; a<mesh->mNumBones; ++a) {
+		for(unsigned int a=0; a<mesh->mNumBones; ++a) {
 			const aiBone* bone = mesh->mBones[a];
             
 			// find the corresponding node by again looking recursively through the node hierarchy for the same name
@@ -366,11 +366,11 @@ void ofxAssimpModelLoader::updateBones() {
 			modelMeshes[i].animatedNorm.assign(modelMeshes[i].animatedNorm.size(),0);
 		}
 		// loop through all vertex weights of all bones
-		for(int a=0; a<mesh->mNumBones; ++a) {
+		for(unsigned int a=0; a<mesh->mNumBones; ++a) {
 			const aiBone* bone = mesh->mBones[a];
 			const aiMatrix4x4& posTrafo = boneMatrices[a];
             
-			for(int b=0; b<bone->mNumWeights; ++b) {
+			for(unsigned int b=0; b<bone->mNumWeights; ++b) {
 				const aiVertexWeight& weight = bone->mWeights[b];
                 
 				size_t vertexId = weight.mVertexId;
@@ -381,7 +381,7 @@ void ofxAssimpModelLoader::updateBones() {
 			if(mesh->HasNormals()){
 				// 3x3 matrix, contains the bone matrix without the translation, only with rotation and possibly scaling
 				aiMatrix3x3 normTrafo = aiMatrix3x3( posTrafo);
-				for(int b=0; b<bone->mNumWeights; ++b) {
+				for(unsigned int b=0; b<bone->mNumWeights; ++b) {
 					const aiVertexWeight& weight = bone->mWeights[b];
 					size_t vertexId = weight.mVertexId;
                     
@@ -435,37 +435,37 @@ ofxAssimpAnimation & ofxAssimpModelLoader::getAnimation(int animationIndex) {
 }
 
 void ofxAssimpModelLoader::playAllAnimations() {
-    for(int i=0; i<animations.size(); i++) {
+    for(unsigned int i=0; i<animations.size(); i++) {
         animations[i].play();
     }
 }
 
 void ofxAssimpModelLoader::stopAllAnimations() {
-    for(int i=0; i<animations.size(); i++) {
+    for(unsigned int i=0; i<animations.size(); i++) {
         animations[i].stop();
     }
 }
 
 void ofxAssimpModelLoader::resetAllAnimations() {
-    for(int i=0; i<animations.size(); i++) {
+    for(unsigned int i=0; i<animations.size(); i++) {
         animations[i].reset();
     }
 }
 
 void ofxAssimpModelLoader::setPausedForAllAnimations(bool pause) {
-    for(int i=0; i<animations.size(); i++) {
+    for(unsigned int i=0; i<animations.size(); i++) {
         animations[i].setPaused(pause);
     }
 }
 
 void ofxAssimpModelLoader::setLoopStateForAllAnimations(ofLoopType state) {
-    for(int i=0; i<animations.size(); i++) {
+    for(unsigned int i=0; i<animations.size(); i++) {
         animations[i].setLoopState(state);
     }
 }
 
 void ofxAssimpModelLoader::setPositionForAllAnimations(float position) {
-    for(int i=0; i<animations.size(); i++) {
+    for(unsigned int i=0; i<animations.size(); i++) {
         animations[i].setPosition(position);
     }
 }
@@ -646,7 +646,7 @@ void ofxAssimpModelLoader::draw(ofPolyRenderMode renderType) {
     ofPushMatrix();
     ofMultMatrix(modelMatrix);
     
-    for(int i=0; i<modelMeshes.size(); i++) {
+    for(unsigned int i=0; i<modelMeshes.size(); i++) {
         ofxAssimpMeshHelper & mesh = modelMeshes[i];
         
         ofPushMatrix();
@@ -881,7 +881,7 @@ int ofxAssimpModelLoader::getNumRotations(){
 
 //-------------------------------------------
 ofPoint ofxAssimpModelLoader::getRotationAxis(int which){
-	if(rotAxis.size() > which){
+	if((int)rotAxis.size() > which){
 		return rotAxis[which];
 	}else{
 		return ofPoint();
@@ -890,7 +890,7 @@ ofPoint ofxAssimpModelLoader::getRotationAxis(int which){
 
 //-------------------------------------------
 float ofxAssimpModelLoader::getRotationAngle(int which){
-	if(rotAngle.size() > which){
+	if((int)rotAngle.size() > which){
 		return rotAngle[which];
 	}else{
 		return 0.0;
